requirements.txt
Flask==2.3.2
tensorflow==2.14.0
numpy==1.25.0
Pillow==10.1.0
gunicorn==21.2.0
